Moved files + changed to grid-layout
